Fix test262 #741 P22/P25/P30
- TypedArray(object): elements are now coerced via ToNumber (ToPrimitive), so a
  throwing valueOf/toString propagates. JSTypedArray no longer hardcodes
  DoubleValue to NaN, which also makes Number(new Int8Array([5])) === 5.
- Array.prototype.slice: the per-element read no longer truncates the source
  index to 32 bits, so array-likes with length up to 2^53-1 slice correctly
  (added a 64-bit TryGetArrayLikeElement overload).
- Function own-key order: the JSFunction constructors install "length" and
  "name" before "prototype", so getOwnPropertyNames yields
  [length, name, prototype, ...] for classes, functions, and native functions.

+++ Broiler.JS/Broiler.JavaScript.BuiltIns/Array/JSArrayPrototype.Utility.cs
@@ -110,7 +110,7 @@ public partial class JSArray
 
         for (long sourceIndex = actualStart; sourceIndex < actualEnd; sourceIndex++)
         {
-            if (!TryGetArrayLikeElement(@this, (uint)sourceIndex, out var value))
+            if (!TryGetArrayLikeElement(@this, sourceIndex, out var value))
             {
                 resultIndex++;
                 continue;
